ResponseEntity 응답추가

+++ chapter3/src/main/kotlin/com/microservices/chapter3/CustomerController.kt
@@ -1,6 +1,8 @@
 package com.microservices.chapter3
 
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
 @RestController
@@ -9,7 +11,11 @@ class CustomerController {
     private lateinit var customerService: CustomerService
 
     @RequestMapping(value = ["/customer/{id}"], method = [RequestMethod.GET])
-    fun getCustomer(@PathVariable id: Int) = customerService.getCustomer(id)
+    fun getCustomer(@PathVariable id: Int): ResponseEntity<Customer> {
+        val customer = customerService.getCustomer(id)
+        val status = if (customer == null ) HttpStatus.NOT_FOUND else HttpStatus.OK
+        return ResponseEntity(customer, status)
+    }
 
     @RequestMapping(value = ["/customers"], method = [RequestMethod.GET])
     fun getCustomers(@RequestParam(required = false, defaultValue = "") nameFilter: String) =
@@ -17,17 +23,30 @@ class CustomerController {
 
 
     @RequestMapping(value = ["/customer"], method = [RequestMethod.POST])
-    fun createCustomer(@RequestBody customer: Customer) {
+    fun createCustomer(@RequestBody customer: Customer): ResponseEntity<Unit> {
         customerService.createCustomer(customer)
+        return ResponseEntity(HttpStatus.CREATED)
     }
 
     @RequestMapping(value = ["/customer/{id}"], method = [RequestMethod.DELETE])
-    fun deleteCustomer(@PathVariable id: Int) {
-        customerService.deleteCustomer(id)
+    fun deleteCustomer(@PathVariable id: Int): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+        if (customerService.getCustomer(id) != null) {
+            customerService.deleteCustomer(id)
+            status = HttpStatus.OK
+        }
+        return ResponseEntity(status)
     }
 
     @RequestMapping(value = ["/customer/{id}"], method = [RequestMethod.PUT])
-    fun updateCustomer(@PathVariable id: Int, @RequestBody customer: Customer) {
-        customerService.updateCustomer(id, customer)
+    fun updateCustomer(@PathVariable id: Int, @RequestBody customer: Customer): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+
+        if (customerService.getCustomer(id) != null) {
+            customerService.updateCustomer(id, customer)
+            status = HttpStatus.ACCEPTED
+        }
+
+        return ResponseEntity(status)
     }
 }
